Add files via upload
Article that inspired me was made by Dabao Zhang, who created this coefficient of determination based on the mean and variance functions. I replicated its coefficient and others previously proposed, comparing them to each other via generation from different distributions. These are the result.

Abstract:
The coefficient of determination, a.k.a. R square, is well-defined in linear regression models, and measures the proportion of variation in the dependent variable explained by the predictors included in the model. To extend it for generalized linear models, we use the variance function to define the total variation of the dependent variable, as well as the remaining variation of the dependent variable after modeling the predictive effects of the independent variables. Unlike other definitions that demand complete specification of the likelihood function, our definition of R2 only needs to know the mean and variance functions, so applicable to more general quasi-models. It is consistent with the classical measure of uncertainty using variance, and reduces to the classical definition of the coefficient of determination when linear regression models are considered.